fix(bug 227 github): we moved the password creation functionality to the team creation
When we create a new team from the ui we were not sending a password so we were creating the
password at the time to deploy (if it not created), but not updating the team element in the
database. So to sdolve it we moved the functionality to create password (if it's not informed) when
we create the team object.